get only required data from db
